+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Validate Canadian postal code format
 - Enforce character limit for location name so it doesn't run off the poster
 - Add QR Alert docs to Portal
 - Add Location Outbreaks History view